remove the useless internal module
