+++ timeline/timeline.cabal
@@ -34,6 +34,5 @@ library
   import:           deps
   hs-source-dirs:   src/
   exposed-modules:  Data.Timeline
-  other-modules:    Data.Timeline.Internal
   default-language: Haskell2010
   ghc-options:      -Wall -fno-warn-unused-do-bind -fno-warn-unused-imports
